Refactoring: moved combineCoordinatesAndqVecs to dataManipulations
